+++ src/test/index.ts
@@ -19,3 +19,4 @@ export * from './PopoverView';
 export * from './DropdownView';
 export * from './InputView';
 export * from './BreadcrumbView';
+export * from './TestPage';
